ai-service/requirements.txt
# Digital HRM AI Service - Python Dependencies
# ==========================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# AI API Clients
openai==1.59.6
anthropic==0.38.0
google-genai==0.8.4

# Redis
redis==5.2.1
aioredis==2.0.2

# HTTP Client
httpx==0.28.1

# PostgreSQL Async Driver
asyncpg==0.30.0

# Retry/Resilience
tenacity==9.0.0

# Token Counting
tiktoken==0.9.0

# Environment Variables
python-dotenv==1.0.1

# File Upload Support
python-multipart==0.0.20

# Async Support
asyncio==3.4.3

# Testing
pytest==8.3.5
pytest-asyncio==0.25.2
pytest-cov==6.1.1
httpx==0.28.1

# Development
black==24.10.0
isort==5.13.2
ruff==0.9.9
mypy==1.14.1

# Optional: For production monitoring
# prometheus-client==0.21.0
# structlog==24.4.0
